Создана сущность MQPresenter
